Wind Boss fight implementation complete (core logic)
-Added full WindBossAI with attack loop system
-Implemented move selection system (multiple abilities)
-Created WindAbilityPack with all core moves: Cyclone, Hurricane Strike, Hurricane Slam, Air Blade, Wind Current, Sky Lift, Eye of the Storm and, Phantom Gust
-Implemented intro cutscene (boss descent + dialogue)
-Implemented death cutscene (UI + dialogue sequence)
-Fixed coroutine restart bug in death cutscene
-Added placeholder visuals for ability telegraphs

+++ Assets/Scripts/IHeal.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IHeal
 {
-    public void Heal(float healAmount);
+    public void Heal(int healAmount);
 }
